Against my wishes I did close to zero coding so far.
I did not count on it, but it seems that to make the backwards steps, I need a few new things, the most important of them being the unzip function.

I thought I would be done with adding new things, but it appears I was wrong.

My biggest problem is that just looking at the things in `SpiralV5Ops` file makes my motivation to go down.

That code is garbage. It is shit.

I guess that Paul Graham was right in his essay that whenever you choose a less powerful language you are making a mistake. But damn it, F# was simply too good for me to resist and spoke too well to my values.

A dynamic language like one of the Lisps would have been too slow for my needs, and a language with strong metaprogramming like D is simply better garbage. When the base is C++ it is hard to improve much upon it.

Functional programming is great. In the last update on the Simulacrum blog, I sang praises of functions, and I still hold the same ideal. It is just that none of the statically typed functional languages meet it.

As great as the type inference in F# is, it is a trap.

And in the dependently typed langauges, instead of allowing metaprogramming via types, they allow theorem proving via types which is an entirely different thing.

I regret it. I had clearly made a mistake.

I picked a language without a strong metaprogramming component and now I am paying for that by making a language with such a component.

The only thing I have to look forward to is that once I have such a language, abstracting other langauges will be quite easy and that includes F#.

I will be able to use Cuda and F# and anything else I desire within the same language. And at that point, I will have what I need.

Forget about this library. I will finish V5, but I won't make any new modules or anything great. I won't have my ideal experimenting environment unless I do the full language.

V5 is only a stepping stone towards that.

Well, enough about that. Let me see if I can figure out how to do the unzip.

I have been unmotivated the entire day today.