Ok, enough for the day. It is 8:25pm for God's sake.
I barely did anything at all until 4pm when I suddenly got into it with the zip and the unzip functions and now I have to pry myself away from the screen.

I managed to reduce the size of zip function by 4x and it fits on the screen now. I am decently sure it is correct, but with the unzip function I am doing evolutionary programming at this point.

I thought I could do it using a generic transpose, but it is obvious I am still missing a few keys on this.

I am actually going to have to look for how to do this thing online and put it through Quickcheck once I am done to make sure it works right.