http://stackoverflow.com/questions/43603903/on-heterogenous-lists-is-it-possible-to-make-the-zip-and-then-unzip-equal-to-th/
This is what I've been trying to do all day today. It is the first time I tried using the FsCheck for anything serious.

I've tightened the code up as much as I could, but I have no idea anymore. Maybe there is some another condition in unzip that I am missing in unzip, but at any rate, I can't see it right now.

For my next move, I'll put these functions into the typechecker and then move on to making the backwards step functions for the launchers.

After that I will be able to do more interesting things with the library.