I am tired and do not feel like continued with this anymore. I improved the array types to make them more composable nad tomorrow I will finish the cuda_module_mapcol_template function.
Today most of my energy was taking up by the two zip functions. I did not manage to make unzip isomorphic with zip, but that does not matter.

I will finish the launcher tomorrow. There are unlikely to be any delays after this as I have all the pieces now.

And after that I will just make the rest of the library.