Look, basically I do not feel like it anymore.
For that shitty 'cuda_module_mapcol_template' function I've been writing those 5 lines for nearly 2h hours already.

I need the parser for this because the syntax is making my eyes spin.

Ghhhhaaaaaa...I did not want to do this. I really did not, not right now. I really want to go into RL. Just what did I do to deserve this? All I want is to have all the pieces together in one place, is that too much to ask?

But the truth of programming is that you can't work hard tomorrow, you must do it today instead. Cutting corners never works.

I keep telling myself that I will write the parser when I've experimented enough, but I think I already passed that point.

Writing the parser for the language would take me around a week or so, but I might as well do it.

It is not like I don't have time for this. I guess in the end, the circumstances are pushing me towards making a full language.