"So let me recap all the patterns here. 1) pat = expr; 2) [pat] = expr; 3) inl name [pat] = expr; 4) inl [pat] → expr; 5) fun name [pat] = expr; 6) fun [pat] → expr;
All of the above will be statements and I will require that every `expr` terminate with an expression.

7) expr;

#2 and #4 will be equivalent to each other and #1...well, I guess it is related to them, though that one is special as it will be evaled on the spot rather than when applied."

The above is an excerpt from my journal. I should have just stopped two hours ago, but then I asked on the Haskell stub about those tuples and now it is 7:30pm so I do not feel like continuing anymore.

All I did today in essence was design the above. Once I have the parser, writting code for the backend and eventually the full langauge will be much easier.

Whenever starting a new thing, I always require some time to rev up, so today it was not surprising that I did not do much.

Well, I'll take it bit by bit. For now I feel like gaming for the rest of the day.