Fixed the infinite loop bug in application by moving it to the top of the calling hierarchy. This simple move required some thought, but took me a bit over half an hour so it was not too bad.
The next on the list will be to put in operators.

One thing that occurred to me is that unless I put in `let` like in F#, I am going to have use `==` for equality like Haskell and C do.

Well, let me have breakfast and I'll gett to it.