I've remove the Unit type in favor of the empty tuple, but as I has trying to put in parenthesless tuples, the kind F# has it occured to me just haw badly I've mismanaged whitespace handling.
So far I've been parsing like `spaces >>. term`, but doing it in an opposite manner such as `term .>> spaces` would be far superior.

The reson being is that in `spaces >>. term` if the first spaces parser succeeds and changes the state, it will cause the loopers to succeed as well, or fail in unexpected ways.

Crap.

Let me do one more version of this.