I am hyped up. This is going really, really well. Right now I am thinking whether I should add indentation sensitive parsing as well.
It would not be a big deal. When I was doing the GVGAI port I was much less experienced than now so it took me a long time to deal with it, but now I do not think that making it would be that big of a deal. The fact that I already got it finished in that port, is a bonus in this as well. I'll have to refactor it a bit, but it should work otherwise.

Besides that I still need to add those operators.

Well, let me stop outside for a bit first. The tuples are done to perfection.