I am out of energy. I am pushing what I have now and will continue tomorrow. I need to think a bit more about how to deal with that infinite loop. Probably, I will build in some kind of loop breaker, not inside the indentation parser, but probably as an independent layer in the parser hierarchy.
In the typechecker, I've been using a stack to good effect, so that seems like a natural method to use here. This would also let me design the parser more freely without having to contort myself with the precise order the things get executed in. With the indentation parser, sometimes they don't get executed at all hence the need for the loop breaker...

I'll think about it overnight, my mind is starting to loop.