I really should have seen this comming, but unfortunately I did not.
I was wondering why

    inl add (a,b) =
        q = 2
        w = 3
        q + w
    0

parses just fine while:

    inl add (a,b) =
        q = a
        w = 3
        q + w
    0

doinks an exception.

After what must be like 30m of thinking it occured to me that it is parsing it like

`q = a w = 3`

Well, I guess now I know why F# uses let statements instead of just `name = expr`.

The loop breaker also have me like an hour of debugging. It was not hard at all making it, but deciding where to put it on the other hand did.

At any rate, now that I understand what the problem is, let me overhaul the syntax for a bit and I will then move on to application.

It won't be long until I am done with this.