I finally got the indentation parser to work, tuples, operators and application. The changes were so simple that it took my breath a way for a moment and even I had to wonder how come I missed this.
It is amazing that I have no need to track indentation in the UserState at all.

It was a lot of struggle getting this going and I think I am going to stop here so I can end the day on a high note.

I tried adding semicolons as well, but it is not that simple. I am going to have to spend some energy to figure it out and I do not have that anymore.

I'll put the finishing touches on this tomorrow and then plug it into the compiler.

Once I do that, writting those kernels will get a lot easier.