"8:50am. I am up. Got the two books.
8:55am. Also I figured out what the problem was with the semicolon parser.

As it turns out, everything. Even though it seems the same at first glance, parsing semicolons in an indentation sensitive language and the one I had in v1b is a completely different thing.

Let me go through two scenarios that completely confused me yesterday while I was low on energy.
let test2 =
    """
    (inl w = 4;
     2)
    """

This one illustrates one of the problems with the way I am doing it now. It returns VV [], but what it actually does is that it actually parses the first statement as it should, sees the semi and then parsers the next expression without caring about the indentation.

Since the program is the same as if I'd wrote
inl w = 4
            2
that is just one statement on the outer level so of course it gets rejected. But the thing is, once it parses the first statement and enters the subexpression, there is literally no way for the parser to turn off and it will consume all the semicolons in its path.

Now comes the second scenario which also returns VV [] for an entirely different reason.
let test2 =
    """
    (inl w = 4
     2)
    """

What happens here is that it parses the first statement and the 4 in the subexpression. Then both the semicolon parsers for the subexpression and for the main level do not find a semi and terminate. The result is that only one statement gets read and then tossed away because statements can't be the last members of a block.

As much as I'd want it to go to the next thing, the only way sepBy can do that is if it gets the separator.

How tricky.

Well, the solution for adding semis is very easy though.

I'll do it right now.
            if i = s.Column then (optional semicolon >>. expr) s
            elif i < s.Column then (semicolon >>. expr) s
            else pzero s

Done. This should work now.

...It does. Also thanks to the second line its behavior now should match F#'s."

The above is an excerpt from my journal. It took me like 5m to make this fix. I figured it out during the night.

With that I am done with the difficult parts of the parser. The rest is merely adding cases and such.

For a few days I think I will take a break to go over the aforementioned books.

Type Driven Development in Idris by Edwin Brady
Pattern Calculus by Barry Jay

There is also the Active Libraries by Todd Veldhuizen thesis which is 270 pages long.

I've been checking the Genesis library for over a month now to get Brady's book and someone finally posted it, so of course I am going to make some time for readings.

Pattern Calculus was posted on the Haskell sub yesterday and it caught my interest and the Active Libraries paper is from the link to Type Systems as Macros on LambdaTheUltimate.

It will take me a while, I'll resume work on the library in a few days. Adios.