I got to around 2/3rds of the Idris book and then got fed up with it.
The parts on theorem proving are not really warming up to me, even six motnhs after I'd last tried dependently typed programming seriously.

Theorem proving as the basis for typed metaprogramming is nonsense and so are the current crop of dependently typed languages. They are just wasting my time.

I should have just kept programming for the last two days. Well, at least I did not waste an entire month and I did get in two hours just now.

Either way, the language is starting to shape up.

Amongst other things, I rewrote the first map module and not only is it half the size of the previous one, but it was also a lot easier to write.

It is remarkable.

The three or so days that I've spent writing the parser have been very well spent indeed.

What I will do tomorrow is spend some time on adding the missing cases and taking advantage of the things I did on the parser to improve the quality of the error messages with positional information.

I could not do this with just syntax sugar that I had previously.

When I said I would be doing this for quite a while three months ago on that /r/types thread I really was not kidding. I've been working on this typechecker for 3 months already.

Let's see if I can 'finish' the library by the end of the month.