http://stackoverflow.com/questions/43755987/how-to-insert-an-indentation-check-into-the-operator-precedence-parser
I posted the question on parsers here. For now, I'll just leave this issue aside. If nothing else, I might repurpose the after string parsers for this sort of task, but right now it is not that important.

Probably, I'll get the reply from the author eventually.

...Let's see...what I'll do next is convert the other modules to the new form and then add the missing cases.

Let me take it nice and smooth to get back in the game. I was eager to get back into programming, but I find myself slacking again today.