"List of things to do from here.
1) Implement named tuples
2) Actual closures
3) Pattern matching on named tuples
4) The :: patterns
5) Redo arrays so as to take advantage of named closures and name filtering in methods.
6) Also mail the author of Fparsec so I get a reply on the issue of the the precedence parser. It does not seem like I will get an answer at this rate."

That course on Ruby did not give me any new insights at all, but it did give me an excuse to turn off for a bit while my brains works out how to do all of the above.

I'll start now and do a bit for the next hour or two or so and aim to complete the above tomorrow.

Once I have the above I'll think a bit on how to do modules.