This is too much of a pain in the ass, I do not feel like it. I should not have tried it all at once
1) Implement named tuples
2) Actual closures
3) Pattern matching on named tuples

Tomorrow I will go at it from the start and just focus on the above three. Leave the arrays alone for now.

...Damn, this is a pain in the ass, but really no other language speaks to my values as well as the one I am doing right now. I have to do it.

I might have messed up and put too much stuff in the typechecker. I need to reconsider it from the start. Maybe there is a better way. There is not way the core should be a 1000 lines long.

I get the feeling that it should be simpler somehow.