Done with introducing named tuples and pattern matching on them. Also I extended the parser a bit and will introduce typecase tomorrow. I feel like I really have a grip on parsers and how to organize them now. Though I still need to test them though.
I think this is enough for the day.

I am only a step away from finishing this backend.  Just focus on finishing the backend, deal with the library and then get into the Game.

Well, if history is any indication I doubt it will be that easy. The Spiral language will become my lasting legacy despite my recent wavering. But that will be in its own time.