I am done with the typecase and nearly ready to start testing during which I'll take Stephan's advice and look at what the thing compiles to using the dotPeek tool.
https://github.com/stephan-tolksdorf/fparsec/issues/9

Currently, I am debating the merits of extending the precedence parser a little. With regard to using the user state for passing around indentation information, my view is that it is overkill. It would be far better to do what I am doing now. The userstate is usefull in cases such as when I had to write that multidimensional array parser and the there were extreme long term dependencies. In comparison, indentation is much simpler.

My way would be a lot more efficient too - in theory. I am going to have to decompile the thing to find out.

Before I can start, I have an annoying issue right now I am trying to deal with. For some reason the File Explorer crashes when I try to open the Torrent folder.

The latest Bahamut ep is like a straw the break the camel's back for it. I do not understand it.

This is beyond annoying. Stupid piece of crap.