"What a rough night. I spent it all agonizing about functions. Actually, going into my parser with a debugger yesterday was enough to make me depressed when I saw what my code decompiled into.
But I will get over it. Mostly I spent the night thinking about the semantics of first class functions in my own language. Do I allow closures and if so, how do I do it?

I am leaning towards no, do not put in closures, even in the full language.

In 95% of cases when I use functions, they would be covered just fine using standard inls. The rest are simple recursive cases like loops. Where I would need closures is where I do things like this in F#.

let tag =
    let mutable i = 0
    fun () -> i <- i+1; i

Clearly, I'd need closures for this kind of case, or would I? I can return inlineables from inlineables just fine, so the above case is already pretty much covered just fine in my language and I would not need to allocate a closure to do it. I am getting the piece of abstraction that I want, namely the variable hidding pretty much for free, and in the most efficient way possible.

I think I will go with this design choice as adding closures will require me to add another pass into the language and it won't make it particulrly more expressive either.

There is an another aspect of functions that needs to be considered. In the language they are mostly being ferried across the type system, meaning they are not trully first class.

Trully first class functions can be stored in contains and later retrieved and passed into the FFI and such. Moreover, using them in the CPS style to accumulate arguments would require them to be instantiated on the term level similarly to the other two use cases.

The times I needed to put functions into containers has been rare, but the times I did it, it has been extremely useful.

Right now I have an example of instantiating function in the CubBlockReduce case, but that thing there is just a hack. I need to generalize it and give it first class support.

I am going to add a special syntax for ApplyType to the language, namely the asterisk operator. When applied on a term, it will instantiate a function on the term level, allowing it to be put into containers and be passed into the FFI and such.

There are some limitations.

I thought for quite a while last night how to get around and I have no choice, by necessity what I am going to do is going to have to be simple and the term level functions themselves are going to have to be simple. Meaning, they won't be allowed to curry or return type level stuff like Inlineables or Methods from them. There is simply no way around this.

The few times I've put functions in containers, they've been extremely simple. I am talking about `unit -> unit` and `float32 -> unit`. And the example I've used yesterday which was int -> int -> bool, could be uncurried to (int, int) -> bool. Alternatively, I could type apply it twice and it would work just as well too, but then I'd get potentially two function calls and two closures instead of one."

Apart from that, I considered turning the parser statements into `let inline`, but I do not feel like performance testing right now. That is for when the program gets bigger and parser performance actually becomes an issue.

What is next on the list?

One thing I've forgotten to do is Mset. I'll add it to the parser promptly as an operator of the lowest precedence. I'll also need to to extend the after parser parser to do indentation checking. After that, I guess I will add ApplyType.