Done with mset. When I said I was done with the day, I played Bayonetta for an hour and then got the urge to do some programming.
For the first time I am testing recursive typechecking in the main typechecker and to my surpise it diverged.

At first I thought it might be because of If hosting or something else that I've overlooked, but tracing it, it looks like the issue is in the closure conversion as that is where it diverges.

Did not count on that.

I'll fix this tomorrow as it is 9:05pm right now.