The fibonacci example with the Y combinator works!
I was not sure if this would be the case but it is.

Beautiful. The Y combinator itself might seem like an academic construct, but I've used the pattern in various guises over and over now and not being able to do it would've been a major blow to me.

Being able to do this was probably the most important thing to me.

Now the only issue is dealing with lambdas produced in recursive functions. If I can deal with that, I will have an language that can exceed all other languages.

If I can clear that condition I will have everything needed to turn this in to a fully fledged one.