I got up 5h ealier than usual today and spent the first 2.5 just now hunting down the typechecking bug. It turns out I forgot to make the arguments types a part of the key in the method dictionary. I did not forget to do this in the reality algorithm prototype.
I still have to fix the closure conversion, but before I do that, I need to think a little how to put in actual closures into my language. The Cuda ones will not have any implicit arguments and will be standard methods or inlineables, but I am going to have to make them in order to cut the recursive chains in the typechecker.

I know what I want to do, but now the question is how to add it with the least amount of friction.