Done with the new typechecker version. Unless I decide to put in explicit rather than the in-place kind of substitution, this will be the last change to the way the language handles functions.
Having the alternative substitution would have allowed the typechecker to not crash on a wider range of programs, but the ones it does crash right now I'd rather explicitly disallow anyway.

I think the way I am using substitution now would be faster anyway.

A lot of thought went into this naming scheme, but in the end it was rather easy to implement and ulike the sceme in the previous version now I can pass environment in the function types directly.

The way this language has been designed has been long in making and today I've finally decided to firmly commit myself to the CPS style.

Try as I might, I do not think it is possible to implement type level closures in this language. I would not say it is completely impossible, but it would be exceedingly difficult to figure a scheme for propagating free variables from closures not in lexical scope.

Methods were always restricted from returning from methods, but today I've decided for the first time to make it a firm rule, even in the full language.

As a hard rule, the language will be firmly focused on making use of continuation passing, but not so much on continuation returning.

...

There were some crazy bugs that I discovered today, like having the variable pushing function grabbing the env from the top level instead via an argument. I had to stop this commit halfway to go hunting.

Also, I redid the renaming part at least 3 times now and I am not completely sure that what I am using now to assign tags to functions is precise...no, not it is not. I just realized how. I have no choice, tomorrow I am going to make a pre-typechecking pass to tag each and every function uniquely. Assigning them as I go along is not working out for me. At the very least, I did figure out what I need to enable passing of environment through the type system. Once I put in TypeApply, I will be able typecheck deeply recursive functions and actually have enough to break the language out of its original purpose of being a Cuda backend. I will be able to do so using term level closures.

And the type level functions which are to be used in 99% of all cases anyway will be pure efficient expressivity.

I'll do it tomorrow, I had enough for today.

Honestly, I should have done the prepass today, but why do it in half an hour when I can be lazy and spend hours fiddling with hacks instead?