As it turns out I am already 500 LOCs in. I did some very major restructuring, but it is true that I wrote the previous version well, so I was able to reuse much of the code.
Refactoring is a breeze in F#. It if this typechecker was only about the quantity of the code, I'd have been done with it in half a week.

I think I will call it a day with this.

I can do this. With today's surge of inspiration I have enough to make the full language and it won't be too hard to make it reality. All the pieces are here now, I just need to put them together.

The way it is coming together is going devastatingly well. If I can program consistently at this level and not get stuck, the Singularity is not out of my reach.