Almost done with the new language core. The only thing missing are arrays.
Weighting at 670 LOCs it feels incredibly sleek and well built now. Code of this quality is trully worty of being the base for Spiral Language.

I started this wanting to make a Cuda backend, but the end result is a language that is both incredibly expressive - it has all the better aspects of impredicative polymorphism - and efficient due to automatic monomorhisation and unboxing and unneed argument elimination which I am pretty much getting for free because of how the language is designed.

This insight of yesterday that I do not need to feed the functions any arguments was the trigger that crystalized my thoughts in the correct direction.

Even though it is 6pm, I think I will put in some overtime and deal with arrays. For two months I had the design where I had special cases for them and even tracked their sizes at the type level, but with the way all the pieces are coming together, I will the duplicate size elimination pretty much for free.

All that code and the special cases for the Arrays are just dead weight now.

After I am done with arrays, in terms of features, in order to really make this a full language I will need to make an interpreter for it and use that to implement macros. I do not really need algebraic datatypes for them, named tuples are enough.

It won't be too hard, the base I have now is incredibly solid, but that particular feature is not a high priority one. I think I will defer it until the time comes to rewrite the Spiral AD library in this language.