"Really, I am such a goof. I figured out how to do free variable propagation.         let env = Set.fold (fun s k ->             match Map.tryFind k d.env with             | Some v -> Map.add k v s             | None -> failwithf "Variable %s not bound." k) Map.empty (expr_free_variables expr)
The above is spew in more ways than one. Not only did I completely forget about the existence of Map.filter, but just like for TyV, I really need to run the expr_free_variable on tthe subfunctions as well.

This is a problem as I do not feel like having to run such a huge calculation that could make multiple passes over the entire tree every time I use call_as_function. I considered instead running that algorithm each time I find a Inlineable, but that would turn my simple pass algorithm into an n one where n is the number of Inlineables.

It took me hours to think of this, but what I could do instead is run the used variable thing just a single time and augment those Inlineables with the list of free variables. Then during typechecking, I could just filter the damn environment based on that before returning a FunctionT in a variable!"

The above is an excerpt from my journal. This particular optimization that I just did will be important the efficiency and the quality of the typechecking. One thing I've also done apart from some renaming is eliminate backing up of the typechecker state in active patterns. That particular operation is simply too expensive.

It will be the programmer's responsibility for making sure that the expressions it passes into the annotation and the function cases do not push any statements.

Also I've fixed the TyMethodCall case so it it takes in the relevant method_tag. Codegen is going to need that.

With this, I can stand back for a bit and admire my handiwork.

Last night I wanted to start work on arrays, but at that point I've been at it for 12h and it was really too much for me. As today shows, going at it with a fresh mind brings significant benefits.

The way the typechecker does free variable related optimizations now is absolutely perfect. There is not a friction in sight.

I guess now can finally turn my attention to arrays. The way Haskell overloads function using the typeclass mechanism is brilliant. I am going to do something similar for tuples.

Right now if they are applied they would just give an error, but I can overload the application mechanism to call a function with the tuple and the argument to it instead.

I'll do something simple like that and use it hijack the indexing mechanism that way.