Done with adding Mset and replacing CudaExpr with strings in the two pattern matching cases.
I think it is time I take a break here. Currently the typechecker is at 710 LOCs. Tomorrow I will work on the parser for a bit again - I'll have to redo the active pattern case and then I'll get back into the codegen.

It won't be hard. The hard part was figuring out this renaming scheme and how to integrate everything with free variable propagation and the zero argument functions trick. The parser was hard to make, but unlike typechecking there was never a doubt in my mind how to do it so it went by quickly. I might be done with parsing a codegen in a few hours tomorrow.