I finally solved the issue that was plaguing me today and the way I plugged on_method_call into the free_variable function is quite brilliant.
As can be expected, it is a breeze to change functional architectures. I have to admit I was surpised today to find out that free_variable propagation did not work as I expected it to. And I am still not sure if two passes are all that is needed for tuple deforestation to converge, but I'll find out if it is not eventually.

At any rate, the thing about there being no friction still applies

Today half my day was taken by this and the rest by the parser.

Maybe I still have some time to take a look at codegen?