This is unbelievable. But those type errors of weeks ago that took me hours to get were definitely an indication that my architecture is not completely well.
Edit: Holy shit, I will never do that again. Never again use higher order functions when mutually recursive blocks will suffice. Never.

I was a damn genius when I spent days making that convoluted structure, grappling with type errors, but here once I started, I did it all in 30m.

And I did apply_closure to boot. It should work now, and I do not even have to add anything special to the codegen to get it to print those methods.

And actually saying that it cost me days is probably understimating it. If I count of the costs of having a convoluted architecture, it probably me a cost a week or more in thinking time.

It is good that I cleared it up now.