The language officially has term level functions. If I was not compiling to Cuda and F# instead, they would be first class functions instead of function pointers.
There are still some things left to work out, but with this, I can move on to bringing back the Cuda functions. That is brilliance.

Last time the function for CubBlockReduce was quite hacky, this time it will be quite a bit principled.

And today I really learned my lesson about using higher order functions where mutually recursive blocks should be used instead.

Good programming is so elegant and subtle. One day I will master it.

As is usual at this point in the day, I've spent all my energy and it is time to rest and recharge for the next battle.

Let me talk a bit about games for a change of pace since I am using these commits as an impromptu journal.

Just yesterday I finally tried Nier Automata for the first time and I can't even change the resolution without it crashing.

In contrast to Bayonetta where I am halfway done with a Pure Platinum run on Normal and no problems at all, Nier's port is incredibly shoddily done.

It will probably take a while before all the bugs get ironed out for the PC version. It is clear the game still needs a few more months in the development. It took forever to crack too.

I can wait.

It is generally a pity that once I spend all of my energy on programming, I do not have that much left over for entertaiment. And for every day that passes, my PC gets further and further behind the curve. During my trading days, that was kind of horror story as I only upgraded my rig two years ago when it was 8 years old.

This might not be bad depending on who the person is, but I was spending 12h in per day in front of the computer then just like now.

It is like a neverending fall, back then when I chased the dream not of easy money, but of defeating the world with my own two hands.

This time the dream is different - I will never again trade, but take by force and terror, and my hands will no longer be gripping a weapon, but a lever with which to move the world.

Though I think of myself as a good programmer, my real talent is not programming, but something else. If I can manifest it, I can stop the fall and win this game.