Today was rather unusual in that instead of stopping at 7pm like I had intended, I simply could not leave that shoddy function alone and kept channeling my will into simplifying it.
In the end I did figure it out.
    and apply_type d (env,core as fun_key) args_ty on_fail ret =
        let args =
            let f x = TyType x |> make_tyv_and_push_inv d
            match args_ty with
            | VVT(l,n) -> TyVV(List.map f l, args_ty)
            | x -> f x
        let f = FunctionT fun_key |> TyType |> T
        let g = FunctionT(env,("",[S " ",meth_memo (ap f (V " "))])) |> TyType
        apply tev d g args on_fail ret

This is what took me like 4 hours of focused effort to write. Simple right. The T case which I've been reluctant to add at first was completely trivial to add and only required a change in two places.

The above is still missing a closure_guard and I will simplify the memoize functions greatly.

Those crappy hacks that I have now are completely unnecessary.