I think I will call it a day here.
Today was strange for me as I was fully focused as I just went through the source code for Fparsec. It is incredibly mutable and imperative in nature, it was almost as if I was reading C code.

There are some significant code duplication and abstraction issues throughout the library, but for all I know it might be written like this simply for performance.

There is no an iota of code inside here that I feel like reusing.

Rewriting the the parser so it passed indentation through the userstate is an option, but not the one I want. What is important here is elegance of design. Once I have my own language, I will rewrite the parser in it where it will in fact be optimal.

Though I was going through the source, I did not read the code much. Instead, as is usual for me, I took the chance to tune out and think about the parallels between D and my own language.

Andrei's talk made me realize something I'd forgotten. D in fact does not have pattern matching, but in my own language everything that they would use a macro, I am using the pattern matcher.

And the pattern matcher, much like F#'s is very rigid and non-extensible.

Everything in it is baked into the language.

If I could instead build them from inside the language I would be able to unlock a explosion of expressivity.

I'd drive the friction in the language down to zero and complete the foundation for the full Spiral language.

At any rate, this parser thing is putting a damper on me. And as I nearing completion of this stage, I want to look into whether I can improve the design of Spiral in this direction.