Unfortunately backin up the parser in the operators function does nothing. I really will have to rewrite the thing.
Yesterday in addition to going through the Fparsec source, I made sure to study Dlang and Pattern Calculus again. Dlang has interesting things, but as a language I consider it too inelegant and bloated for my tastes. And I got a bit futher in Pattern Calculus today.

The author is seriously not speaking to my values In part 3 there are finally some code examples, and the language has some interesting things like pattern matching on values, but I can't say that there is anything new for me to take there.

As disappointing as my studies are my brain more than made up the slack in terms of ideas yesterday.

"1) Convert the language to CPS and use a special Result type at the end to make sure I don't miss calling a continuation.
2) Clean up stack on failure using the same trick I am already using in the if case if the false branch was not popped.
3) Rework pattern matching so that only switch statements can be applied and this time, do not compile them all away. Instead I can use this to turn named tuples into super intelligent classes."

Doing 1 and 2 will allow me to do extensible active patterns and is well worth doing just for that, but with regard to #3, I am not that sure about it. I think I will avoid it. The thing I have in mind right now regarding it is just too thorny and the embrionic idea I had yesterday is just not good.

Just extending active patterns to their full potential should be enough to satisfy my extensibility needs.

But damn this parser. I really wish the Stephan did it right the first time. No doubt, the version I will make will be 10x slower than his, but it does not really matter.

After I finish the new active patterns, the only things missing to make Spiral a full language will be adding discriminated unions and a F# codegen.

But that is in the future. Right now, let me just put my nose to the grindstone and deal with the precedence parser and then the transpose and zip functions.